requirement.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
reportlab==4.0.0
python-dotenv==1.0.0
pytest==7.4.0
